chore: centralize the createEither utility function - clean up console.log and dead comments
