+++ src/Either.ts
@@ -1,5 +1,3 @@
-import { createArrayProxy } from "./utils"
-
 /**
  * Either type represents a data structure that encapsulates a successful result or an Error.
  * It wraps the result of a Promise in an object, making it easier to handle errors by returning
@@ -16,32 +14,3 @@ export type Either<T> =
       result: T
       error: undefined
     } & [undefined, T])
-
-export const createEither = <T>({
-  result,
-  error,
-}:
-  | {
-      error: Error
-      result: undefined
-    }
-  | {
-      error: undefined
-      result: T
-    }): Either<T> => {
-      if (error) {
-        const array: [Error, undefined] = [error, undefined]
-        // const obj = Object.create(array)
-        const obj = {} as any
-        obj.error = error
-        obj.result = undefined
-        return createArrayProxy<T>(obj, array)
-      }
-      const array: [undefined, T] = [undefined, result]
-      // const obj = Object.create(array)
-      const obj = {} as any
-      obj.error = undefined
-      obj.result = result
-      return createArrayProxy<T>(obj, array)
-    }
-
