+++ src/go/Either.ts
@@ -1,4 +1,3 @@
-import { createArrayProxy } from "../utils"
 /**
  * Either type represents a data structure that encapsulates a successful result or an Error.
  * It wraps the result of a Promise in an object, making it easier to handle errors by returning
@@ -15,36 +14,3 @@ export type Either<T> =
       result: T
       error: undefined
     } & [T, undefined])
-
-
-
-
-
-
-export const createEither = <T>({
-  result,
-  error,
-}:
-  | {
-      error: Error
-      result: undefined
-    }
-  | {
-      error: undefined
-      result: T
-    }): Either<T> => {
-  if (error) {
-    const array: [undefined, Error] = [undefined, error]
-    // const obj = Object.create(array)
-    const obj = {} as any
-    obj.error = error
-    obj.result = undefined
-    return createArrayProxy<T>(obj, array)
-  }
-  const array: [T, undefined] = [result, undefined]
-  // const obj = Object.create(array)
-  const obj = {} as any
-  obj.error = undefined
-  obj.result = result
-  return createArrayProxy<T>(obj, array)
-}
